fix: cast duration to int to avoid float precision issues
Ensure the duration is cast to an integer to prevent potential float precision issues when converting milliseconds to seconds

+++ test3.py
@@ -322,7 +322,7 @@ class PancakeRobotNode(Node):
         for frequency, duration in notes:
             note = AudioNote()
             note.frequency = frequency
-            note.max_runtime = Duration(sec=duration/1000)
+            note.max_runtime = Duration(sec=int(duration/1000))
             note_msg.notes.append(note)
         self.audio_publisher.publish(note_msg)
